integrated server invite/join flow
